refactor(undo): group per-domain recorders under undo/recorders
Move node.py and flow.py into an undo/recorders/ subpackage so the mechanism
(core, recording, manager) is separated from per-domain reversal knowledge, and
new domains have an obvious home. Pure move plus import/docstring updates.
