+++ src/griptape_nodes/retained_mode/managers/undo/recorders/node.py
